requirements.txt
appdirs==1.4.4
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
blinker==1.5
CacheControl==0.12.11
cachetools==5.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
cssselect==1.2.0
cssutils==2.6.0
dnspython==2.3.0
docker==6.0.1
dominate==2.7.0
email-validator==1.3.1
firebase-admin==6.0.1
firestore==0.0.8
Flask==2.2.2
Flask-Bootstrap4==4.0.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
gcloud==0.17.0
google-api-core==2.11.0
google-api-python-client==2.73.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.9.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
greenlet==2.0.1
grpcio==1.51.1
grpcio-status==1.51.1
h11==0.14.0
httplib2==0.21.0
idna==3.4
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
jws==0.1.3
lxml==4.9.2
MarkupSafe==2.1.1
MouseInfo==0.1.3
msgpack==1.0.4
oauth2client==3.0.0
outcome==1.2.0
packaging==22.0
premailer==3.10.0
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.21
pycryptodome==3.4.3
PyGetWindow==0.0.9
PyJWT==2.6.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
Pyrebase==3.0.27
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-jwt==2.0.1
pytweening==1.0.4
pywin32==305
requests==2.28.2
requests-toolbelt==0.7.0
robotframework==6.0.1
robotframework-pythonlibcore==4.0.0
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary @ git+https://github.com/robotframework/SeleniumLibrary.git@24c732b1628420ab8fd016424489ba1d63667503
rsa==4.9
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.13
visitor==0.1.3
webdrivermanager==0.10.0
websocket-client==1.4.2
Werkzeug==2.2.2
wsproto==1.2.0
WTForms==3.0.1
yagmail==0.15.293